Created the layout for experience page
Added sections for each of my job roles I've had in the past, detailing the company, job title and dates of employment. I have added a paragraph section for adding more details, but have kept it to placeholder text for now. I also made the page interactive by adding a transition that increases the section scale on hover.