Set up basic project navigation and added pages for individual projects
I have added pages for all of the main projects in my portfolio and also linked them to the main project page. The main project page has been set up to be a navigation between my projects and can give a small view into each before navigating to the page. For now, only the title and photo appear in the grid, with the photo being a placeholder, but I will add in the real content soon.