Created basic project page layout with the creation of the first project page
Made the first project showcase page, which will serve as a basic layout for the rest of the pages. This page specifically shows the project Farm to Fork, and my involvement in it. For now I consider this page to be fully complete in terms of content - there are no placeholders. Although I may go back to add some sections to give some more detail (more images/ footage from the completed project).